requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.3
openai==0.27.5
pandas==2.0.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.3.23
requests==2.29.0
scipy==1.10.1
six==1.16.0
tiktoken==0.3.3
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.9.2
